Fixed numbering throughout
Reverted internals to use 0-4, made sure cards and hints only _display_ 1-5. (This was how I was lead to believe it would work originally. This means the large numbers on the cards are actually affected (they weren't before), and it also means hints aren't kinda broken (they were before, become some of the logic was using 1-5 while the rest of it still expected 0-4).