Add extra padding to bottom of Mergers to account for Android bottom bar
Ideally the fix for this would be to avoid using 100vh, which doesn't account for the bottom bar in Android. But in order to do that I'd probably change max-height to 100%, and in order for that to work we'd have to set the height of the surrounding div in GameDarkSublayout to also be 100%: https://github.com/freeboardgames/FreeBoardGames.org/blob/master/web/src/gamesShared/components/fbg/GameDarkSublayout.tsx#L125

I tried that out on Android and it appears to work for Mergers on Android at least, but that's more cross-cutting of a change than I want to make right now. Going to stick with this since it works fine.